requirements.txt
flask>=2.0.0
python-dotenv>=0.19.0
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
chromadb>=0.4.22
requests>=2.26.0
pypdf
tiktoken
werkzeug
google-generativeai>=0.3.0
gunicorn==21.2.0
